Refactor code to create 'json_list' instead of 'json_hash'.
Although a hash/dict seems to follow the standard appearance
of json data examples online, it does not follow the intentions
of the task parameters. Compiling 3-main.py with 'json_hash'
results in an error from trying to append to a hash. This in-
dicates that the data structure truly needs to be a list.

Compiling 3-main.py with 'json_list' succeeds and using a list
with save_to_file/json.dump does not appear to pose an issue.