Modify Car to accept dictionary/hash input to constructor.
Define hash of attributes (hash var) according to whether kwargs
or args has values. If kwargs has len > 0, arguments are being
passed to constructor with keywords (and most likely not as a
hash). But if kwargs is len < 1 and args has len > 0, then assume
that attributes are being passed as a single hash.