Write several possibilities for structuring cases in main.py.
Mostly as an exercise to think through which option is best.
Going with main-1.py because it is the most natural and readable
to the person writing the code.

The fact that checking for improper input comes last might cause
several unnecessary computations but they are very simple com-
putations with light cost. So readability is being prioritized.