Handle case where no last name is passed for student.
Use len(argv) to discriminate between cases where both last name
and first name are passed, and wheres only one name is given.

In the case where len(argv) == 6, do not specify a first_name
value in the call to create.
This avoids a list index error accessing argv when creating the
record.

Currently the print option prints an extra space before the name of
the student, if only one name is passed. This may need correction
in the future.