Fix delete_record method in main.
A record can be deleted either by directly specifying which record
in the table to delete, as in

   <table>.delete().where(<table>.id==X)

or by selecting/getting a recording, storing it into a variable
and using

    <record>.delete_instance.

Since the record being deleted must be selected or retrieved prior
to deleting in order to print it, the 2nd method makes most sense
for the specs of the program. Use the 2nd method.

Also remove type of exception for except clause, since an opera-
tional error does not occur when a record does exist, but some
exception does.